Rank support added to ivmeasurer.

+++ testall.bat
@@ -1,9 +1,9 @@
 @echo off
 setlocal EnableDelayedExpansion
-rd /S /Q venv
+if exist venv  rd /S /Q venv
 python -m venv venv
 venv\Scripts\python -m pip install --upgrade pip
-venv\Scripts\python -m pip install -r requirements.txt
+venv\Scripts\python -m pip install -r requirements.txt --upgrade
 echo --- Run all tests ---
 for /D %%f in (epcore\*) do (
 echo Test %%f
